fix: Fix u16 overflow by changing type to int

+++ include/progress.h
@@ -14,7 +14,7 @@
 
 typedef struct {
         const char      *label;
-        u16             tty_width;
+        int             tty_width;
         u64             total;
         u64             current;
         struct timespec started;
